move comments, add testutilv
