+++ packages/language-support/src/tests/formatting/testutilv2.ts
@@ -1,0 +1,17 @@
+import { formatQuery } from '../../formatting/formattingv2';
+import { standardizeQuery } from '../../formatting/standardizer';
+
+export function verifyFormatting(query: string, expected: string): void {
+  const formatted = formatQuery(query);
+  expect(formatted).toEqual(expected);
+  const queryStandardized = standardizeQuery(query);
+  const formattedStandardized = standardizeQuery(formatted);
+  if (formattedStandardized !== queryStandardized) {
+    throw new Error(
+      `Standardized query does not match standardized formatted query`,
+    );
+  }
+  // Idempotency check
+  const formattedTwice = formatQuery(formatted);
+  expect(formattedTwice).toEqual(formatted);
+}
